% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_users_uncertainty.R
\name{adjust_users_uncertainty}
\alias{adjust_users_uncertainty}
\title{Adjusting quantity data to reflect users with uncertainty.}
\usage{
adjust_users_uncertainty(
  sectors_reporting_input,
  reporting_rates_df,
  baseline_users,
  commodities_table,
  method_continuation_df,
  user_input_adjustment_table
)
}
\arguments{
\item{sectors_reporting_input}{A table of values indicating if a sector is fully/partial/non-reporting.}

\item{reporting_rates_df}{A table of reporting rates for your chosen ss_type}

\item{baseline_users}{A table of baseline users calculated for the years prior to your chosen ss_type data.}

\item{commodities_table}{A table of commodities for your chosen service statistic type}

\item{method_continuation_df}{A table of method continuation rates for long term methods.}

\item{user_input_adjustment_table}{A table indicating whether a method should be adjusted or not.}
}
\value{
A table returning the samples of users for each method each year adjusted for the missing private sector.
}
\description{
Adjusting quantity data to reflect users with uncertainty.
}
