% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_users.R
\name{baseline_users}
\alias{baseline_users}
\title{Calculating the Baseline Users}
\usage{
baseline_users(
  commodities_table,
  service_stats_type,
  scale_up_table,
  method_contination_rates
)
}
\arguments{
\item{commodities_table}{A table of commodities for your chosen service statistic type}

\item{service_stats_type}{The type of service statistics chosen.}

\item{scale_up_table}{Recoded table using scale up actions of service statistics.}

\item{method_contination_rates}{Method continutaion rates for long term methods.}
}
\value{
A table calculating the baseline users for the years prior to those included in the data.
}
\description{
Calculating the Baseline Users
}
