% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_emu_from_users.R
\name{calculate_emu_from_users}
\alias{calculate_emu_from_users}
\title{Calculating EMU incl. and excl. condoms using population data and adjusted users}
\usage{
calculate_emu_from_users(
  pop_data,
  users_data,
  fixed_users_data,
  service_stats_type,
  reporting_rate_data,
  method_summary
)
}
\arguments{
\item{pop_data}{A table of population data}

\item{users_data}{A table of adjusted users totalled for the LAPM and STM methods - calculated using samples.}

\item{fixed_users_data}{A table of adjusted users totalled for the LAPM and STM methods - calculated using standard fixed method.}

\item{service_stats_type}{service statistics type}

\item{reporting_rate_data}{A dataframe of reporting rates by year}

\item{method_summary}{Summarise by method or not (TRUE/FALSE)}
}
\value{
A table of calculated EMU data
}
\description{
Calculating EMU incl. and excl. condoms using population data and adjusted users
}
