% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shiny_tools_info.R
\name{get_shiny_tools_info}
\alias{get_shiny_tools_info}
\title{Get tools information from SS-to-EMU Excel spreadsheet for the ss2emu shiny app}
\usage{
get_shiny_tools_info(country_file_path, shiny_input_type)
}
\arguments{
\item{country_file_path}{Location of the country's SS-to-EMU Excel spreadsheet}

\item{shiny_input_type}{Chosen service statistics type}
}
\description{
This function reads an Excel spreadsheet containing tools information
and returns the relevant information.
}
