% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tools_info.R
\name{get_tools_info}
\alias{get_tools_info}
\title{Get tools information from SS-to-EMU Excel spreadsheet}
\usage{
get_tools_info(country_file_path)
}
\arguments{
\item{country_file_path}{Location of the country's SS-to-EMU Excel spreadsheet}
}
\description{
This function reads an Excel spreadsheet containing tools information
and returns the relevant information.
}
