% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emu_data.R
\name{plot_emu_data}
\alias{plot_emu_data}
\title{Plotting EMU data at the level and rate of change.}
\usage{
plot_emu_data(emu_data, mcpr_data)
}
\arguments{
\item{emu_data}{A table of EMU data for one chosen type}

\item{mcpr_data}{A table of mCPR run data}
}
\value{
A plot of EMU data
}
\description{
Plotting EMU data at the level and rate of change.
}
